fix(thumos): extract kinit pure logic into host-testable kinit_plan (#528)
Refs #528
